requirements.txt
boto3==1.9.111
pillow==5.4.1
numpy==1.16.2
